% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire-seq.R
\name{splitToCells}
\alias{splitToCells}
\title{Create a SplitDataFrameList}
\usage{
splitToCells(df, field, ids = df[[field]], universe = sort(unique(ids)))
}
\arguments{
\item{df}{A data frame of repertoire data where each row corresponds to a detected sequence.}

\item{field}{A string specifying the field of \code{df} that contains the identity of the cell.}

\item{ids}{A character vector of length equal to \code{nrow(df)} specifying the identity of the cell for each sequence.}

\item{universe}{Character vector of all possible cell identities.}
}
\value{
A \linkS4class{SplitDataFrameList} containing a per-cell perspective of repertoire data.
}
\description{
Create a SplitDataFrameList to store repertoire data.
Each cell corresponds to one DataFrame, while each row of that DataFrame corresponds to a sequence detected in the cell.
}
\details{
Setting \code{universe} is useful to ensure that the output object is of the same length as,
say, the number of columns in a \linkS4class{SingleCellExperiment} object containing expression data.
}
\examples{
df <- data.frame(
    cell.id=sample(LETTERS, 30, replace=TRUE),
    v_gene=sample(c("TRAV1", "TRAV2", "TRAV3"), 30, replace=TRUE),
    j_gene=sample(c("TRAJ4", "TRAJ5", "TRAV6"), 30, replace=TRUE)
)

splitToCells(df, field="cell.id")

splitToCells(df, field="cell.id", universe=c(LETTERS, "AA"))
    
}
\author{
Aaron Lun
}
