% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeClonalExpansion.R
\name{summarizeClonalExpansion}
\alias{summarizeClonalExpansion}
\title{Summarize clonal expansion}
\usage{
summarizeClonalExpansion(
  counts,
  use.mean = TRUE,
  use.gini = TRUE,
  use.top = c(5, 20, 100)
)
}
\arguments{
\item{counts}{A list of integer vectors such as that produced by \code{\link{countCellsPerClonotype}}.
Each vector corresponds to a group of cells and contains the number of cells for each clonotype in that group.}

\item{use.mean}{Logical scalar indicating whether to report the mean number of cells per clonotype.}

\item{use.gini}{Logical scalar indicating whether to report the Gini index.}

\item{use.top}{Integer vector specifying the number of clonotypes to use to compute the top percentage.}
}
\value{
A \linkS4class{DataFrame} with one row per group in \code{counts},
containing summary statistics on the clonal expansion in that group.
}
\description{
Generate some summary statistics for clonal expansion based on the number of cells per clonotype.
}
\details{
If \code{use.mean=TRUE}, the output will contain the numeric \code{"mean"} column,
containing the average number of cells per clonotype computed across all clonotypes for a given group.
Larger values indicate that cells are concentrated within a small number of clonotypes.

If \code{use.gini=TRUE}, the output will contain the numeric \code{"gini"} column,
containing the Ginit index for clonotype diversity in each group.
Larger values indicate that cells are concentrated within a small number of clonotypes.

If \code{use.top} is specified, the output will contain the numeric \code{"topX"} columns,
containing the proportion of cells assigned to the top X clonotypes (where X is the value of each \code{use.top} entry).
Larger values indicate that cells are concentrated within a small number of clonotypes.
}
\examples{
df <- data.frame(
    cell.id=sample(LETTERS, 30, replace=TRUE),
    clonotype=sample(paste0("clonotype_", 1:5), 30, replace=TRUE)
)

y <- splitToCells(df, field="cell.id")
out <- countCellsPerClonotype(y, "clonotype",
   group=sample(3, length(y), replace=TRUE))

summarizeClonalExpansion(out)

}
\author{
Aaron Lun
}
