% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeClonotypeCounts.R
\name{summarizeClonotypeCounts}
\alias{summarizeClonotypeCounts}
\title{Summarize clonotype counts}
\usage{
summarizeClonotypeCounts(
  counts,
  use.mean = TRUE,
  use.gini = TRUE,
  use.top = c(5, 20, 100),
  use.hill = 0:2,
  downsample = TRUE,
  down.ncells = NULL
)
}
\arguments{
\item{counts}{A list of integer vectors such as that produced by \code{\link{countCellsPerClonotype}}.
Each vector corresponds to a group of cells and contains the number of cells for each clonotype in that group.}

\item{use.mean}{Logical scalar indicating whether to report the mean number of cells per clonotype.}

\item{use.gini}{Logical scalar indicating whether to report the Gini index.}

\item{use.top}{Integer vector specifying the number of clonotypes to use to compute the top percentage.}

\item{use.hill}{Integer scalar specifying the orders to use to compute Hill numbers.}

\item{downsample}{Logical scalar indicating whether downsampling should be performed.}

\item{down.ncells}{Integer scalar indicating the number of cells to downsample each group to.
Defaults to the smallest number of sequence-containing cells across all levels in \code{group}.}
}
\value{
A \linkS4class{DataFrame} with one row per group in \code{counts},
containing summary statistics on the clonal expansion in that group.
}
\description{
Generate some summary statistics for clonal expansion based on the number of cells per clonotype.
}
\details{
If \code{use.mean=TRUE}, the output will contain the numeric \code{"mean"} column,
containing the average number of cells per clonotype computed across all clonotypes for a given group.
Larger values indicate that cells are concentrated within a small number of clonotypes.

If \code{use.gini=TRUE}, the output will contain the numeric \code{"gini"} column,
containing the Gini index for clonotype diversity in each group.
Larger values indicate that cells are concentrated within a small number of clonotypes.

If \code{use.top} is specified, the output will contain the numeric \code{"topX"} columns,
containing the proportion of cells assigned to the top X clonotypes (where X is the value of each \code{use.top} entry).
Larger values indicate that cells are concentrated within a small number of clonotypes.

If \code{use.hill} is specified, the output will contain numeric \code{"hillX"} columns containing the Hill numbers.
\code{"hill0"} is simply the number of observed clonotypes (i.e., species richness)
while \code{"hill1"} and \code{"hill2"} quantify the evenness of the distribution of cells across clonotypes.
(\code{"hill2"} differs from \code{"hill1"} in that the former gives more weight to dominant clonotypes.)

See \code{\link{countClonotypesPerCell}} for why downsampling is turned on by default.
}
\examples{
df <- data.frame(
    cell.id=sample(LETTERS, 30, replace=TRUE),
    clonotype=sample(paste0("clonotype_", 1:5), 30, replace=TRUE)
)

y <- splitToCells(df, field="cell.id")
out <- countCellsPerClonotype(y, "clonotype",
   group=sample(3, length(y), replace=TRUE))

summarizeClonotypeCounts(out)

}
\author{
Aaron Lun
}
