% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeGeneComboCounts.R
\name{summarizeGeneComboCounts}
\alias{summarizeGeneComboCounts}
\title{Summarize gene combination counts}
\usage{
summarizeGeneComboCounts(
  counts,
  use.gini = TRUE,
  use.top = c(5, 20, 100),
  use.hill = 0:2,
  downsample = TRUE,
  down.ncells = NULL
)
}
\arguments{
\item{counts}{A \linkS4class{SummarizedExperiment} containing cell counts for each gene combination (row) and group (column),
such as that produced by \code{\link{countCellsPerGeneCombo}}.
Alternatively, a count matrix containing the same information.}

\item{use.gini}{Logical scalar indicating whether to report the Gini index.}

\item{use.top}{Integer vector specifying the number of clonotypes to use to compute the top percentage.}

\item{use.hill}{Integer scalar specifying the orders to use to compute Hill numbers.}

\item{downsample}{Logical scalar indicating whether downsampling should be performed.}

\item{down.ncells}{Integer scalar indicating the number of cells to downsample each group to.
Defaults to the smallest number of sequence-containing cells across all levels in \code{group}.}
}
\value{
A \linkS4class{DataFrame} with one row per group in \code{counts},
containing summary statistics on the diversity of gene expression in that group.
}
\description{
Generate some summary statistics for the gene diversity, based on the number of cells expressing each gene combination.
}
\details{
If \code{use.gini=TRUE}, the output will contain the numeric \code{"gini"} column,
containing the Gini index for gene combination diversity in each group.
Larger values indicate that a small number of gene combinations are expressed in many cells.

If \code{use.top} is specified, the output will contain the numeric \code{"topX"} columns,
containing the proportion of cells expressing the top X gene combinations (where X is the value of each \code{use.top} entry).
Larger values indicate that a small number of gene combinations are expressed in many cells.

If \code{use.hill} is specified, the output will contain numeric \code{"hillX"} columns containing the Hill numbers.
\code{"hill0"} is simply the number of observed gene combinations (i.e., species richness)
while \code{"hill1"} and \code{"hill2"} quantify the evenness of the distribution of cells across gene combinations.
(\code{"hill2"} differs from \code{"hill1"} in that the former gives more weight to dominant gene combinations.)

See \code{\link{countCellsPerGeneCombo}} for an explanation of why downsampling is turned on by default.
}
\examples{
df <- data.frame(
    cell.id=sample(LETTERS, 30, replace=TRUE),
    v_gene=sample(c("TRAV1", "TRAV2", "TRAV3"), 30, replace=TRUE),
    j_gene=sample(c("TRAJ4", "TRAJ5", "TRAV6"), 30, replace=TRUE)
)

y <- splitDataFrameByCell(df, field="cell.id")
out <- countCellsPerGeneCombo(y, c("v_gene", "j_gene"),
   group=sample(10, length(y), replace=TRUE))

summarizeGeneComboCounts(out)

}
\author{
Aaron Lun
}
