% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiPerReadPerCell.R
\name{umiPerReadPerCell}
\alias{umiPerReadPerCell}
\title{UMIs per read per cell}
\usage{
umiPerReadPerCell(x, read.field, umi.field)
}
\arguments{
\item{x}{A \linkS4class{SplitDataFrameList} where each DataFrame is a cell and each row is a sequence.}

\item{read.field}{String containing the name of the column containing the read count data.}

\item{umi.field}{String containing the name of the column containing the UMI count data.}
}
\value{
A numeric vector containing the ratio of UMIs to reads for each cell.
}
\description{
Compute the UMI-to-read ratio for each cell.
}
\details{
This function is designed to evaluate the degree of redundancy in the read coverage of each UMI.
Low values indicate that the reads are highly redundant such that little can be gained from further sequencing.

Note that, in repertoire data, the definition of \dQuote{low} is somewhat different from usual.
This is because only deeply sequenced transcripts will survive the assembly and annotation process,
such that the reported sequences are likely to be biased towards very low UMI:read ratios.
Values below 0.01 are typical.
}
\examples{
df <- data.frame(
    cell.id=sample(LETTERS, 30, replace=TRUE),
    v_gene=sample(c("TRAV1", "TRAV2", "TRAV3"), 30, replace=TRUE),
    j_gene=sample(c("TRAJ4", "TRAJ5", "TRAV6"), 30, replace=TRUE),
    reads=rnbinom(30, mu=20, size=0.5)
    umis=rnbinom(30, mu=2, size=1)
)

y <- splitToCells(df, field="cell.id")
umiPerReadPerCell(y, "reads", "umis")

}
\author{
Aaron Lun
}
